+++ src/event.rs
@@ -8,6 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+// TODO: Only because of `derive_more` macros, try to remove on next
+//       `derive_more` upgrade.
+#![allow(clippy::use_self)]
+
 //! Key occurrences in a lifecycle of [Cucumber] execution.
 //!
 //! The top-level enum here is [`Cucumber`].
@@ -396,6 +400,16 @@ impl<World> Clone for Step<World> {
 /// [`Step`]: gherkin::Step
 #[derive(Clone, Debug, Display, Error, From)]
 pub enum StepError {
+    /// [`Step`] doesn't match any [`Regex`].
+    ///
+    /// It's emitted whenever a [`Step::Skipped`] event cannot be tolerated
+    /// (such as when [`fail_on_skipped()`] is used).
+    ///
+    /// [`Regex`]: regex::Regex
+    /// [`fail_on_skipped()`]: crate::WriterExt::fail_on_skipped()
+    #[display(fmt = "Step doesn't match any function")]
+    NotFound,
+
     /// [`Step`] matches multiple [`Regex`]es.
     ///
     /// [`Regex`]: regex::Regex