+++ CHANGELOG.md
@@ -6,6 +6,25 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 
 
 
+## [0.18.0] · 2022-??-??
+[0.18.0]: /../../tree/v0.18.0
+
+[Diff](/../../compare/v0.17.0...v0.18.0) | [Milestone](/../../milestone/21)
+
+### BC Breaks
+
+- Added `NotFound` variant to `event::StepError`. ([#250])
+
+### Fixed
+
+- Not panicking on `fail_on_skipped()` with retries. ([#250], [#249])
+
+[#249]: /../../issues/249
+[#250]: /../../pull/250
+
+
+
+
 ## [0.17.0] · 2022-11-23
 [0.17.0]: /../../tree/v0.17.0